pyproject.toml
[tool.poetry]
name = "tlo-devs-commons"
version = "0.1.0"
description = ""
authors = ["ThaRising <kochbe.ber@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"

[tool.poetry.dev-dependencies]
django = "^3.1.2"
google-cloud-storage = "^1.31.2"
sqlalchemy = "^1.3.19"
pydantic = "^1.6.1"
pytest = "^6.1.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
